Corrigindo tipos primitivos e ordem de precedencia

+++ Exercicios/exer_03.py
@@ -1,7 +1,7 @@
 # Desenvolva um programa que leia as duas notas de um aluno, calcule e mostre a sua média
 
-nota1 = int(input('Digite a sua primeira nota: '))
-nota2 = int(input('Digite a sua segunda nota: '))
+nota1 = float(input('Digite a sua primeira nota: '))
+nota2 = float(input('Digite a sua segunda nota: '))
 
 media = (nota1 + nota2) / 2
